+++ cmd/mailchain/internal/serve.go
@@ -21,7 +21,6 @@ import (
 	"github.com/mailchain/mailchain/cmd/mailchain/config"
 	"github.com/mailchain/mailchain/cmd/mailchain/config/defaults"
 	"github.com/mailchain/mailchain/cmd/mailchain/internal/rest/handlers"
-	"github.com/mailchain/mailchain/internal/pkg/http/rest/ethereum/address/messages"
 	"github.com/mailchain/mailchain/internal/pkg/http/rest/ethereum/address/publickey"
 	"github.com/mailchain/mailchain/internal/pkg/http/rest/ethereum/messages/send"
 	"github.com/mailchain/mailchain/internal/pkg/http/rest/messages/read"
@@ -78,7 +77,7 @@ func CreateRouter(cmd *cobra.Command) (http.Handler, error) {
 	r.HandleFunc("/api/ethereum/{network}/address/{address:[-0-9a-zA-Z]+}/public-key", publickey.GetPublicKey(pubKeyFinders)).Methods("GET")
 	r.HandleFunc(
 		"/api/ethereum/{network}/address/{address:[-0-9a-zA-Z]+}/messages",
-		messages.GetMessages(mailboxStore, receivers, keystore, deriveKeyOptions)).Methods("GET")
+		handlers.GetMessages(mailboxStore, receivers, keystore, deriveKeyOptions)).Methods("GET")
 	r.HandleFunc("/api/ethereum/{network}/messages/send", send.SendMessage(sentStorage, senders, keystore, deriveKeyOptions)).Methods("POST")
 	r.HandleFunc("/api/messages/{message_id}/read", read.GetRead(mailboxStore)).Methods("GET")
 	r.HandleFunc("/api/messages/{message_id}/read", read.PutRead(mailboxStore)).Methods("PUT")
